Move the augment to main
